Closes FB case 193
subtemplate: check that if the hash-table access worked, any new pvars are compatible with the old ones on which the cache is based (in the sense of "no new non-#f positions").

Also check that all derived variables from the same xᵢ have the same shape (i.e. don't filter out derived variables for this check).